add api-walkthrough project
