+++ modules/ROOT/examples/csharp/api-walkthrough.sln
@@ -1,0 +1,21 @@
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.26430.14
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "api-walkthrough", "api-walkthrough\api-walkthrough.csproj", "{C16E6840-531A-4B77-8A8E-4530E00D9E10}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{C16E6840-531A-4B77-8A8E-4530E00D9E10}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C16E6840-531A-4B77-8A8E-4530E00D9E10}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C16E6840-531A-4B77-8A8E-4530E00D9E10}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C16E6840-531A-4B77-8A8E-4530E00D9E10}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
