+++ modules/ROOT/examples/csharp/api-walkthrough/api-walkthrough/api-walkthrough.csproj
@@ -1,0 +1,27 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp2.0</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Couchbase.Lite.Enterprise" Version="2.1.0-db001" />
+    <PackageReference Include="SkiaSharp.Components.Layout" Version="0.6.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <None Update="avatar.jpg">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
+    <None Update="travel-sample.cblite2\db.sqlite3">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
+    <None Update="travel-sample.cblite2\db.sqlite3-shm">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
+    <None Update="travel-sample.cblite2\db.sqlite3-wal">
+      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+    </None>
+  </ItemGroup>
+</Project>
